+++ CHANGELOG.md
@@ -2,8 +2,15 @@
 
 All notable changes to the "code-analyzer" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+## 1.0.0
 
-## [Unreleased]
+- Initial release;
 
-- Initial release
+## 2.0.1
+
+- Allow users to set their own OpenAI API key;
+
+## 3.0.0
+
+- Allow users to set their language;
+- UI/UX improvements;